+++ tests/Unit/EventSub/TwitchySharp.EventSub.Webhooks.Tests.Unit/ProcessStubs.cs
@@ -1,4 +1,6 @@
-﻿using TwitchySharp.EventSub.Serialization;
+﻿using System.Collections.Immutable;
+using TwitchySharp.EventSub.Notifications;
+using TwitchySharp.EventSub.Serialization;
 using TwitchySharp.EventSub.Webhooks.Functional;
 using TwitchySharp.Infrastructure.Functional;
 
@@ -49,4 +51,21 @@ internal class ProcessStubs
 
     public static ProcessWebhookRequest StubProcess { get; }
         = (request, ct) => ValueTask.FromResult<Validation<WebhookRequestContent>>(CreateFakeContent(request.Content));
+
+    public static EventSubSubscription FakeSubscription { get; } = new()
+    {
+        Id = new("f1c2a387-161a-49f9-a165-0f21d7a4e1c4"),
+        Status = EventSubSubscriptionStatus.Enabled,
+        Type = new("channel.follow"),
+        Version = new("1"),
+        Cost = 1,
+        Condition = new Dictionary<string, object>() { { "broadcaster_user_id", "12826" } }.ToImmutableDictionary(),
+        CreatedAt = DateTimeOffset.Parse("2019-11-16T10:11:12.634234626Z"),
+        Transport = new() { Method = EventSubTransportMethod.Webhook, Callback = new("https://example.com/webhooks/callback") }
+    };
+}
+
+public record StubEventSubNotification : IEventSubNotification
+{
+    public EventSubSubscription Subscription => ProcessStubs.FakeSubscription;
 }
